[TSK] REF Web Client [REF] separando as resposabiblidade do webclient como o interceptor e o requisicoes da transaction
