+++ lib/http/interceptors/logging_interceptor.dart
@@ -1,0 +1,21 @@
+import 'package:http_interceptor/http_interceptor.dart';
+
+class LoggingInterceptor implements InterceptorContract {
+  @override
+  Future<RequestData> interceptRequest({RequestData data}) async {
+    print('Request');
+    print('url: ${data.url}');
+    print('headers: ${data.headers}');
+    print('body: ${data.body}');
+    return data;
+  }
+
+  @override
+  Future<ResponseData> interceptResponse({ResponseData data}) async {
+    print('Response');
+    print('status code: ${data.statusCode}');
+    print('headers: ${data.headers}');
+    print('body: ${data.body}');
+    return data;
+  }
+}
